Adds styleguide-html wrapper around example code
The _styleguide_block.erb appears to have been missing the
styleguide-html wrapper around the styleguide-code pre element.
The missing wrapper results in the code example being poorly
laid out and visually unappealing.

Adds the wrapper around the element. This corrects the display
and looks more as expected from the example screenshots on
http://warpspire.com/kss/styleguides/